对用户进行改装

+++ guns/guns-admin/src/main/java/com/stylefeng/guns/modular/system/model/UserInfo.java
@@ -51,7 +51,7 @@ public class UserInfo extends Model<UserInfo> {
     /**
      * 钱
      */
-    private BigDecimal money;
+    private int money;
     /**
      * 登陆IP
      */
